join static lib
