+++ DAL2 dont touch/DAL2/message.h
@@ -6,31 +6,31 @@
 #include <QString>
 #include <QTime>
 #include "statement.h"
+namespace SQLDAL {
+	class Message : public Statement {
+	public:
+		Message(std::shared_ptr<Connection> connection);
+		Message();
 
-class Message : public Statement {
- public:
-  Message(std::shared_ptr<Connection> connection);
-  Message();
+		unsigned int		id;
+		unsigned int		chat_id;
+		unsigned int		owner_id;
+		QString			data;
+		QDate				date;
+		QTime				time;
+		bool				status;
 
-  unsigned int		id;
-  unsigned int		chat_id;
-  unsigned int		owner_id;
-  QString			data;
-  QDate				date;
-  QTime				time;
-  bool				status;
+		QVector<Message>	GetMessages();  // SelectAllInfoAboutOneChat
+		void			    AddNewMessage();
+		void              DeleteMessage();
+		void			    UpdateMessage();
 
-  QVector<Message>	GetMessages();  // SelectAllInfoAboutOneChat
-  void			    AddNewMessage();
-  void              DeleteMessage();
-  void			    UpdateMessage();
-
- private:
-  QString UpdateQuery();
-  QString SelectQuery();
-  QString InsertQuery();
-  QString DeleteQuery();
-  QString CreateQueryCountOfMessages();
-};
-
+	private:
+		QString UpdateQuery();
+		QString SelectQuery();
+		QString InsertQuery();
+		QString DeleteQuery();
+		QString CreateQueryCountOfMessages();
+	};
+}
 #endif  // MESSAGE_H
