+++ DAL2 dont touch/DAL2/statement.cpp
@@ -1,24 +1,25 @@
 #include "statement.h"
+namespace SQLDAL {
+	Statement::Statement(std::shared_ptr<Connection> connection) {
+		connection_ = connection;
+		SetQuery(connection_->GetDB());
+	}
 
-Statement::Statement(std::shared_ptr<Connection> connection) {
-  connection_ = connection;
-  SetQuery(connection_->GetDB());
-}
+	Statement::~Statement() {}
 
-Statement::~Statement() {}
+	void Statement::SetQuery(QSqlDatabase database) {
+		query_ = QSqlQuery(database);
+	}
 
-void Statement::SetQuery(QSqlDatabase database) {
-  query_ = QSqlQuery(database);
-}
+	void Statement::ExectuteQuery(QString query) {
+		query_.prepare(query);
+		if (!query_.exec()) {
+			ErrorInfo();
+		}
+	}
 
-void Statement::ExectuteQuery(QString query) {
-  query_.prepare(query);
-  if (!query_.exec()) {
-    ErrorInfo();
-  }
-}
-
-void Statement::ErrorInfo() {
-  qDebug() << query_.lastError().databaseText();
-  qDebug() << query_.lastError().driverText();
+	void Statement::ErrorInfo() {
+		qDebug() << query_.lastError().databaseText();
+		qDebug() << query_.lastError().driverText();
+	}
 }
