+++ DAL2 dont touch/DAL2/statement.h
@@ -4,27 +4,27 @@
 #include <qsqlrecord.h>
 #include <QString>
 #include <memory>
-#include "connection.h"
+#include "Connection/connect.h"
 #include "unitofwork.h"
 
 #define CLIENT_DB "../DAL2 dont touch/DAL2/Client"
 #define SERVER_DB "../DAL2 dont touch/DAL2/Server"
+namespace SQLDAL {
+	class Statement {
+	public:
+		Statement(std::shared_ptr<Connection> connection);
+		Statement() = default;
+		~Statement();
 
-class Statement {
- public:
-  Statement(std::shared_ptr<Connection> connection);
-  Statement() = default;
-  ~Statement();
-
-  protected:
+	protected:
 
 
-  void SetQuery(QSqlDatabase database);
-  void ExectuteQuery(QString query);
-  void ErrorInfo();
+		void SetQuery(QSqlDatabase database);
+		void ExectuteQuery(QString query);
+		void ErrorInfo();
 
-  std::shared_ptr<Connection> connection_;
-  QSqlQuery query_;
-};
-
+		std::shared_ptr<Connection> connection_;
+		QSqlQuery query_;
+	};
+}
 #endif  // !STATEMENT_H
