+++ DAL2 dont touch/DAL2/unitofwork.cpp
@@ -1,15 +1,16 @@
 #include "unitofwork.h"
+namespace SQLDAL {
+	std::atomic<unsigned int> UnitOfWork::connection_number_ = 0;
 
-std::atomic<unsigned int> UnitOfWork::connection_number_ = 0;
+	UnitOfWork::UnitOfWork() {
+		connection_ = std::make_shared<Connection>(GenerateNewConnection());
+	}
 
-UnitOfWork::UnitOfWork() {
-  connection_ = std::make_shared<Connection>(GenerateNewConnection());
-}
+	UnitOfWork::~UnitOfWork() {
+		connection_->Close();
+	}
 
-UnitOfWork::~UnitOfWork() { 
-  connection_->Close(); 
-}
-
-QString UnitOfWork::GenerateNewConnection() {
-  return QString("Connection number:" + QString::number(connection_number_++));
+	QString UnitOfWork::GenerateNewConnection() {
+		return QString("Connection number:" + QString::number(connection_number_++));
+	}
 }
