+++ DAL2 dont touch/DAL2/user.h
@@ -1,49 +1,48 @@
 #ifndef USER_H
 #define USER_H
 #include <QString>
-#include "connection.h"
 #include "statement.h"
+namespace SQLDAL {
+	class User : public Statement {
+	public:
+		User(std::shared_ptr<Connection> connection);
+		~User();
 
-class User : public Statement {
- public:
-  User(std::shared_ptr<Connection> connection);
-  ~User();
-
-  unsigned int	id;
-  QString		login;
-  QString		password;
-  QString		ip;
-  unsigned int	port;
+		unsigned int	id;
+		QString		login;
+		QString		password;
+		QString		ip;
+		unsigned int	port;
 
 
-  void			AddNewFriend(const unsigned int& friend_id);
-  QVector<int>	GetFriends();// SelectAllFriendsAboutOneUser
-  void			DeleteFriend(const unsigned int& friend_id);
+		void			AddNewFriend(const unsigned int& friend_id);
+		QVector<int>	GetFriends();// SelectAllFriendsAboutOneUser
+		void			DeleteFriend(const unsigned int& friend_id);
 
-  void			AddNewFriendRequest(const unsigned int& friend_id);
-  QVector<int>	GetFriendsRequest();  // SelectAllFriendsRequestAboutOneUser
-  void			DeleteFriendRequest(const unsigned int& friend_id);
-  
-  void			GetUser();             // Select
-  void			UpdateUser();          // Update
-  void			AddNewUser();          // Insert
-  void			DeleteUser();          // Delete
+		void			AddNewFriendRequest(const unsigned int& friend_id);
+		QVector<int>	GetFriendsRequest();  // SelectAllFriendsRequestAboutOneUser
+		void			DeleteFriendRequest(const unsigned int& friend_id);
 
- private:
-  QString InsertFriendQuery(unsigned int friend_id);
-  QString SelectFriendsQuery();
-  QString DeleteFriendQuery(const unsigned int& friend_id);
+		void			GetUser();             // Select
+		void			UpdateUser();          // Update
+		void			AddNewUser();          // Insert
+		void			DeleteUser();          // Delete
 
-  QString InsertFriendRequestQuery(unsigned int friend_id);
-  QString SelectFriendsRequestQuery();
-  QString DeleteFriendRequestQuery(const unsigned int& friend_id);
-  
-  QString UpdateQuery();
-  QString SelectQuery();
-  QString InsertQuery();
-  QString DeleteQuery();
-  QString CountOfFriends();
-  QString CountOfFriendsRequest();
-};
+	private:
+		QString InsertFriendQuery(unsigned int friend_id);
+		QString SelectFriendsQuery();
+		QString DeleteFriendQuery(const unsigned int& friend_id);
 
+		QString InsertFriendRequestQuery(unsigned int friend_id);
+		QString SelectFriendsRequestQuery();
+		QString DeleteFriendRequestQuery(const unsigned int& friend_id);
+
+		QString UpdateQuery();
+		QString SelectQuery();
+		QString InsertQuery();
+		QString DeleteQuery();
+		QString CountOfFriends();
+		QString CountOfFriendsRequest();
+	};
+}
 #endif  // !USER_H
